refactor(agent-core, cli, desktop): 收口内建宿主工具模型契约
