+++ packages/agent-core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './ports.js';
 export * from './runtime.js';
 export * from './openai/index.js';
+export * from './host-tools.js';
 export * from './mcp/index.js';
